refactor: moved api/trading-pairs/:url_symbol endpoint logic into a service and fixed caching fallback in /api/ticker
