+++ src/app/api/services/TradingPairsServices.ts
@@ -6,15 +6,35 @@ interface TradingPair {
   description: string;
 }
 
-export class TradingPairsService {
-  private readonly apiUrl: string;
+interface TickerData {
+  timestamp: string;
+  open: string;
+  high: string;
+  low: string;
+  last: string;
+  volume: string;
+  vwap: string;
+  bid: string;
+  ask: string;
+  side: string;
+  open_24: string;
+  percent_change_24: string;
+}
 
-  constructor(apiUrl: string = 'https://www.bitstamp.net/api/v2/trading-pairs-info/') {
-    this.apiUrl = apiUrl;
+export class TradingPairsService {
+  private readonly tradingPairsApiUrl: string;
+  private readonly tickerApiBaseUrl: string;
+
+  constructor(
+    tradingPairsApiUrl: string = 'https://www.bitstamp.net/api/v2/trading-pairs-info/',
+    tickerApiBaseUrl: string = 'https://www.bitstamp.net/api/v2/ticker/'
+  ) {
+    this.tradingPairsApiUrl = tradingPairsApiUrl;
+    this.tickerApiBaseUrl = tickerApiBaseUrl;
   }
 
   public async fetchTradingPairs(): Promise<TradingPair[]> {
-    const response = await axios.get<TradingPair[]>(this.apiUrl);
+    const response = await axios.get<TradingPair[]>(this.tradingPairsApiUrl);
     return response.data.map(pair => ({
       name: pair.name,
       url_symbol: pair.url_symbol,
@@ -35,6 +55,11 @@ export class TradingPairsService {
       totalItems: tradingPairs.length,
       items: paginatedPairs,
     };
+  }
+
+  public async fetchTickerData(urlSymbol: string): Promise<TickerData> {
+    const response = await axios.get<TickerData>(`${this.tickerApiBaseUrl}${urlSymbol}/`);
+    return response.data;
   }
 
   public validatePageParam(pageParam: string | null): number {
