+++ src/app/api/trading-pairs/[url_symbol]/route.ts
@@ -1,22 +1,8 @@
 import { NextResponse } from 'next/server';
-import axios from 'axios';
-
-interface TickerData {
-  timestamp: string;
-  open: string;
-  high: string;
-  low: string;
-  last: string;
-  volume: string;
-  vwap: string;
-  bid: string;
-  ask: string;
-  side: string;
-  open_24: string;
-  percent_change_24: string;
-}
+import { TradingPairsService } from '../../services/TradingPairsServices';
 
 export async function GET(request: Request) {
+  const service = new TradingPairsService();
   const { pathname } = new URL(request.url);
   const url_symbol = pathname.split('/').pop();
 
@@ -25,9 +11,8 @@ export async function GET(request: Request) {
   }
 
   try {
-    const response = await axios.get<TickerData>(`https://www.bitstamp.net/api/v2/ticker/${url_symbol}/`);
-
-    return NextResponse.json(response.data);
+    const tickerData = await service.fetchTickerData(url_symbol);
+    return NextResponse.json(tickerData);
   } catch (error) {
     return NextResponse.json({ error: 'Failed to fetch trading pair data' }, { status: 500 });
   }
